InstantDrawableを移動、不要なコードを削除
